requirements.txt
python-dotenv
streamlit
pyyaml
vertexai
tenacity